Don't compress file
